+++ CHANGELOG.md
@@ -5,3 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Changed
+- Instagram handler fixed as of `2019-05-14 01:00:00 GMT+8`
+
+### Added
+- Functions to mitigate Content-Origin Rules